tweetduck.install: uppercase all non-templated variables
Python uses the dollar sign ($) as the symbol that defines strings that are to be templated, which as you can guess would cause problems with PowerShell scripts. The solution is to just capitalize all variables and object fields, as Python's default templating package is case-sensitive, in which otherwise some cases may cause these fields or variables to be improperly replaced.