ludusavi: simplify packaging by including binaries directly
This greatly simplifies how this package installs by removing the need to unzip an archive to get the binaries. Now they are directly bundled into the package and no script needs to run at all! Currently, I haven't removed it, as it needs to remove the binary from the previous installation, but this paves the way to remove it entirely in the future.

Closes https://github.com/Starz0r/ChocolateyPackagingScripts/issues/14